requirements.txt
boto3~=1.35.56
requests~=2.32.3
PyYAML~=6.0.2
pandas~=1.5.3
python-dotenv~=1.0.1